Fix some type warnings
Changes tagdb_add() and atom_ensure_loc_space()/atom_ensure_blk_space()
to store power2gt()'s result directly in an unsigned int instead of
casting it to size_t, matching power2gt()'s actual return type. Casts
two strlen() results to int at their bmfont_measure()/bmfont_draw()
call sites in bmfont-test.c, where those functions take an int length.
In rev-s-block.c, replaces a pointer-to-int cast used for alignment
checking with intptr_t, and includes base/types.h for it.

+++ libraries/databases/tag-db/tag-db.c
@@ -457,10 +457,10 @@ result_t tagdb_add(tagdb_t *db, const unsigned char *name, tagdb_tag_t *ptag)
         goto Failure;
       }
 #else
-      size_t n;
-      void  *newarr;
+      unsigned int n;
+      void        *newarr;
 
-      n = (size_t) power2gt(db->c_allocated);
+      n = power2gt(db->c_allocated);
       if (n < 8)
         n = 8;
